LayoutTests/http/wpt/identity/identitycredentialscontainer-get-hidden.https.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>
    Digital Credential API: get() can't be used with hidden documents.
</title>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body></body>
<script>
    promise_setup(async () => {
        if (document.visibilityState !== "hidden") {
            await new Promise((resolve) => {
                document.onvisibilitychange = resolve;
                testRunner.setPageVisibility("hidden");
            });
        }
    });

    promise_test(async function (test) {
        this.add_cleanup(() => {
            testRunner.setPageVisibility("visible");
        });

        assert_equals(
            document.visibilityState,
            "hidden",
            "now should be hidden"
        );

        await test_driver.bless();

        assert_true(
            navigator.userActivation.isActive,
            "User activation should be active after test_driver.bless()."
        );

        await promise_rejects_dom(
            test,
            "NotAllowedError",
            navigator.identity.get({})
        );
    }, "navigator.identity.get() can't be used with hidden documents.");
</script>
